services/import-service/requirements.txt
# Web framework & server
fastapi==0.115.12         # updated: embeds starlette>=0.40.0 (fixes CVE-2024-47874)
uvicorn==0.34.0           # updated: latest stable

# Data & file parsing
pandas==2.2.3             # updated: minor security patch
openpyxl==3.1.5           # updated: latest stable
xlrd==2.0.1
xlwt==1.3.0

# HTTP & form parsing
python-multipart==0.0.22  # updated: fixes CVE-2024-24762, CVE-2024-53981, CVE-2026-24486
httpx==0.28.1             # updated: latest stable

# Validation & config
pydantic==2.11.3          # updated: latest stable
python-dotenv==1.1.0      # updated: latest stable

# Messaging
kafka-python==2.0.2

# Testing (dev only)
pytest==8.3.5             # updated: latest stable
